# Allocate the FBuffer struct on the stack